+++ factory.py
@@ -225,6 +225,18 @@ class PretrainedHFOpenCLIPFactory(Factory):
 
         image_encoder = HFOpenCLIPImageEncoder(hf_open_clip_model.vision_model)
         image_projector = HFOpenCLIPImageProjector(hf_open_clip_model.visual_projection)
+
+        if self._peft.name == 'lora':
+            config = LoraConfig(r=self._peft.r,
+                                target_modules=self._peft.target_modules,
+                                lora_alpha=(self._peft.r * self._peft.alpha_r_scale),
+                                lora_dropout=self._peft.dropout,
+                                layers_to_transform=self._peft.layers_to_transform
+                               )
+            image_encoder = get_peft_model(image_encoder, config)
+        else:
+            raise TypeError(f'{self._peft.name} is not supported.')
+
         clip = HFOpenCLIP(image_encoder, image_projector, hf_open_clip_model, self._temperature)
 
         if self._mae == 'pixel':
@@ -244,17 +256,6 @@ class PretrainedHFOpenCLIPFactory(Factory):
 
         # [NOTE]: creating model...
         model = MAECLIP(image_encoder, clip, mae, alpha=self._alpha)
-
-        if self._peft.name == 'lora':
-            config = LoraConfig(r=self._peft.r,
-                                target_modules=self._peft.target_modules,
-                                lora_alpha=(self._peft.r * self._peft.alpha_r_scale),
-                                lora_dropout=self._peft.dropout,
-                                layers_to_transform=self._peft.layers_to_transform
-                               )
-            model = get_peft_model(model, config)
-        else:
-            raise TypeError(f'{self._peft.name} is not supported.')
 
         # [NOTE]: all the trainable parameters are requires_grad = False.
         for name, param in model.named_parameters():
